requirements.txt
pillow
streamlit 
torch